fix(init): delegate rust line length ownership Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
