+++ .editorconfig
@@ -11,6 +11,7 @@ max_line_length = 120
 indent_size = 4
 
 [*.rs]
+# Rust line length is handled by rustfmt
 indent_size = 4
 max_line_length = off
