+++ .github/config/flint.toml
@@ -1,6 +1,7 @@
 [settings]
 exclude = ["CHANGELOG.md", "super-linter-versions/**", "tests/cases/**"]
 setup_version = 2
+setup_migration_version = 2
 
 [checks.renovate-deps]
 exclude_managers = ["github-actions", "github-runners", "cargo"]
