--- .textlintrc
@@ -1,8 +1,0 @@
-{
-  "filters": {
-    "comments": true
-  },
-  "rules": {
-    "terminology": true
-  }
-}
